move query to `firebase-queries`
still have a circular dep issue
